\name{hytraj07.lines}
\alias{hytraj07.lines}
\docType{data}

\title{
	HYSPLIT Trajectory Lines 
}

\description{
  This object is an output example from the function \code{\link{Df2SpLines}} 
  using the data set \code{\link{hy.traj2007}} as input, and CRS = "+proj=longlat +datum=NAD83 +no_defs +ellps=GRS80 +towgs84=0,0,0".
}

\usage{data(hytraj07.lines)}

\examples{
# data(hytraj07.lines)
# str(hytraj07.lines)

# library(sp)
# plot(hytraj07.lines)
}

\keyword{datasets}
