\name{hytraj07.linesDf}
\alias{hytraj07.linesDf}
\docType{data}
\title{
	HYSPLIT Trajectory Lines Data Frame 
}
\description{
  This object is an output example from the function \code{\link{Df2SpLinesDf}} 
  using the object \code{\link{hytraj07.lines}} and the data set \code{\link{hy.traj2007}} 
  as input values.
}
\usage{data(hytraj07.linesDf)}

\examples{
## data(hytraj07.linesDf)
## str(hytraj07.linesDf)

## library(sp)
## plot(hytraj07.linesDf)
}
\keyword{datasets}
