\name{ph07.traj.freq}
\alias{ph07.traj.freq}
\docType{data}
\title{
	Pheno 2007 trajectory frequency calculation
}
\description{
  This Object of class RasterLayer was calculated by function \code{\link{RasterizeTraj}}, and
  it represents the trajectory frequency for the SpatialLines defined by \code{\link{hy.traj2007}}
  with resolution grid value of 10000 kilometers.
}
\usage{data(ph07.traj.freq)}

\keyword{datasets}
