\name{worldmap}
\alias{worldmap}
\docType{data}
\title{
  World's Map
}
\description{
  This dataset contains the world's map that is used by function \code{\link{PlotBgMap}} to 
  display geography background.
}
\usage{data("worldmap")}

\details{
  This dataset is a low resolution representation of the world's map.
}

\examples{
  data(worldmap)
}
\keyword{World's Map}
